Cleanup webpack things, move it to netlify-cms and cleanup dependencies
