+++ package.json
@@ -58,7 +58,6 @@
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.0.1",
     "babel-jest": "^23.6.0",
-    "babel-loader": "^8.0.4",
     "babel-plugin-emotion": "^9.2.11",
     "babel-plugin-inline-svg": "^1.0.0",
     "babel-plugin-lodash": "^3.3.4",
@@ -74,7 +73,6 @@
     "enzyme-adapter-react-16": "^1.0.2",
     "eslint": "^5.3.0",
     "eslint-plugin-react": "^7.10.0",
-    "friendly-errors-webpack-plugin": "^1.7.0",
     "http-server": "^0.11.1",
     "jest": "^23.4.0",
     "jest-cli": "^23.4.0",
@@ -92,12 +90,10 @@
     "rollup-plugin-react-svg": "^2.1.0",
     "rollup-plugin-replace": "^2.1.0",
     "start-server-and-test": "^1.7.0",
-    "style-loader": "^0.21.0",
     "stylelint": "^9.4.0",
     "stylelint-config-recommended": "^2.1.0",
     "stylelint-config-styled-components": "^0.1.1",
-    "stylelint-processor-styled-components": "^1.3.1",
-    "svg-inline-loader": "^0.8.0"
+    "stylelint-processor-styled-components": "^1.3.1"
   },
   "workspaces": [
     "packages/*"
